# PERF: unstack 2 the unstackening